+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* JSON serialization will now include `npm`, `componentDir`, and `moduleResolution` fields.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.10.0] - 2018-03-12